backend/pyproject.toml
[project]
name = "securescan"
version = "0.10.3"
description = "Diff-aware security scanning with multi-tool orchestration, deterministic SARIF, and PR-comment rendering."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "Metbcy"}]
keywords = ["security", "sast", "dast", "sbom", "sarif", "ci"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    # typer 0.21+ requires click>=8.2 (uses click.Choice[T] subscript at
    # import time); older typer paired with newer click also breaks. Pin
    # both floors so a fresh `pip install ./backend` in CI resolves to a
    # compatible pair.
    "typer[all]>=0.21.0",
    "click>=8.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
    "jinja2>=3.1.0",
    "defusedxml>=0.7.1",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
pdf = ["weasyprint>=62.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
    "weasyprint>=62.0",
]

[project.urls]
Homepage = "https://github.com/Metbcy/securescan"
Repository = "https://github.com/Metbcy/securescan"
Issues = "https://github.com/Metbcy/securescan/issues"
Changelog = "https://github.com/Metbcy/securescan/blob/main/CHANGELOG.md"

[project.scripts]
securescan = "securescan.cli:app"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "securescan",
    "securescan.api",
    "securescan.scanners",
    "securescan.ai",
]

[tool.setuptools.package-data]
securescan = ["data/**/*", "templates/**/*"]

[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = [
    "venv",
    ".venv",
    "build",
    "dist",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "I",      # isort
    "UP",     # pyupgrade — modernizes typing imports
    "B",      # flake8-bugbear
    "ASYNC",  # async pitfalls
]
# Bandit (S) deliberately NOT in `select`. The codebase's subprocess
# hygiene was independently audited as good (every `create_subprocess_exec`
# uses arg-list passing, no shell=True). Adding S would flag every call
# (S603/S607 in particular) without surfacing a real risk. Reconsider only
# if the codebase ever introduces shell=True.
ignore = [
    "E501",  # line length — let the formatter handle it
    "B008",  # FastAPI uses `= Depends(...)` in defaults; not a real bug
    # ASYNC240 fires on every `Path.exists()`/`os.path.isdir()` call inside
    # `async def`. Fixing the existing 24 sites requires either
    # `aiofiles`/`anyio.Path` (new dep) or `await asyncio.to_thread(...)`
    # wrappers — both are behaviour-changing refactors that don't belong
    # in a lint-bring-up PR. TODO(v0.11): re-enable and migrate.
    "ASYNC240",
]

[tool.ruff.lint.per-file-ignores]
# Tests can do whatever they want.
"tests/**/*.py" = ["B", "ASYNC"]
# CLI module-init re-exports for back-compat — wildcard imports are fine.
"securescan/cli/__init__.py" = ["F401", "F403"]
# main.py must call `load_user_env()` before any module that instantiates
# Settings is imported (Settings reads env at class-body time). The late
# imports below `load_user_env()` are the documented pattern.
"securescan/main.py" = ["E402"]
